refactor: move error handling to domain of patch notes
