--- app/src/main/java/com/daluwi/sc_news/core/error_handling/Error.kt
@@ -1,3 +1,0 @@
-package com.daluwi.sc_news.core.error_handling
-
-sealed interface Error
