+++ app/src/main/java/com/daluwi/sc_news/features/patches/domain/error_handling/Error.kt
@@ -1,0 +1,3 @@
+package com.daluwi.sc_news.features.patches.domain.error_handling
+
+sealed interface Error
